Clean up Dockerfile and Makefile
The previous Dockerfile had some issues which have been
reported
(see: corona-zahlen-landkreis/corona_landkreis_fallzahlen_scraping#62).
Consequently, this commit tries to address those:

Packages which are not needed to run the project will be removed
(e.g.: sudo and git).

The two volumes will be removed, because their purpose remains unclear.
Instead a new volume will be created providing the python-based source code.

In order to preserve file ownership when accessing files
from within the container, a new user will be created with
the same UID and GUI as the host user id.

The Dockerfile and the Makefile have been moved into the
subfolder holding the python-based crawler logic which prevents
Docker from loading the whole project into the build-context
and simultanouesly directly implies what this Dockerfile is dedicated
to (the crawler, not the api).