Fix Makefile run target
The current run target does not properly spawn a subshell which results
in an "invalid mount config" when `docker run` tries to
inteprates the `--volume` parameter. Hence, this commit prevents `make` from
trying to expand the string to a variable in the first place.
(fixes: corona-zahlen-landkreis/corona_landkreis_fallzahlen_scraping#67)